requirements.txt
asgiref==3.8.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Django==5.2
django-redis==5.4.0
djangorestframework==3.16.0
h11==0.16.0
packaging==25.0
pycparser==2.22
PyMySQL==1.1.1
redis==5.2.1
sqlparse==0.5.3
tzdata==2025.2
waitress==3.0.2
whitenoise==6.9.0
